Moved sliceDateRange to parent component to keep ChartContent state agnostic
