+++ apps/google-analytics-4/frontend/src/components/main-app/AnalyticsApp.tsx
@@ -2,12 +2,13 @@ import { useAutoResizer, useFieldValue } from '@contentful/react-apps-toolkit';
 import ChartFooter from 'components/main-app/ChartFooter';
 import ChartHeader from 'components/main-app/ChartHeader';
 import { useEffect, useState } from 'react';
-import { DateRangeType, RunReportResponse } from 'types';
+import { DateRangeType, Row, RunReportResponse } from 'types';
 import { config } from '../../config';
 import ChartContent from './ChartContent';
 
 const AnalyticsApp = () => {
   const [runReportResponse, setRunReportResponse] = useState<RunReportResponse>({} as RunReportResponse);
+  const [pageViewData, setPageViewData] = useState<RunReportResponse>(runReportResponse)
   const [slugValue] = useFieldValue<string>('slug');
   const [dateRange, setDateRange] = useState<DateRangeType>('lastWeek')
   useAutoResizer();
@@ -23,8 +24,29 @@ const AnalyticsApp = () => {
     fetchData();
   }, []);
 
+  useEffect(() => sliceByDateRange(dateRange), [dateRange, runReportResponse])
+
   const handleDateRangeChange = (e: DateRangeType) => {
     setDateRange(e);
+    sliceByDateRange(e)
+  }
+
+  const sliceByDateRange = (dateRange: DateRangeType) => {
+    if (runReportResponse.rowCount) {
+      let newRows: Row[] = [];
+      switch (dateRange) {
+        case 'lastDay':
+          newRows = runReportResponse.rows.slice(0, 2);
+          break;
+        case 'lastWeek':
+          newRows = runReportResponse.rows.slice(0, 7);
+          break;
+        case 'lastMonth':
+          newRows = runReportResponse.rows.slice(0, 28);
+          break;
+      }
+      setPageViewData({ ...runReportResponse, rows: [...newRows] });
+    }
   }
 
   const pageViews =
@@ -43,7 +65,7 @@ const AnalyticsApp = () => {
         metricValue={pageViews || pageViews === 0 ? pageViews.toString() : ''}
         handleChange={handleDateRangeChange}
       />
-      {runReportResponse.rowCount ? <ChartContent pageViewData={runReportResponse} dateRange={dateRange} /> : null}
+      {pageViewData.rowCount ? <ChartContent pageViewData={pageViewData} dateRange={dateRange} /> : null}
       <ChartFooter slugName={slugValue ? slugValue : ''} viewUrl="https://analytics.google.com/" />
     </>
   );
